multiLanguage/requirements.txt
alembic==1.0.5
appnope==0.1.0
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.6.0
autopep8==1.3.5
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==3.0.2
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
charts==0.4.6
click==6.7
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
diff-match-patch==20121119
dlib==19.16.0
docx==0.2.4
dpath==1.4.2
entrypoints==0.2.3
et-xmlfile==1.0.1
face-recognition==1.2.3
face-recognition-models==0.3.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Migrate==2.3.1
Flask-Script==2.0.6
Flask-SocketIO==3.1.1
Flask-SQLAlchemy==2.3.2
hyperlink==18.0.0
idna==2.6
imageio==2.4.1
incremental==17.5.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.2
jieba==0.39
Jinja2==2.10
json-spec==0.10.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kazoo==2.6.1
kiwisolver==1.0.1
labelImg==1.8.1
lxml==4.2.1
Mako==1.0.7
MarkupSafe==1.1.0
matplotlib==2.2.2
mistune==0.8.4
MyQR==2.3.1
mysql-connector==2.1.6
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.4
numpy==1.14.5
opencv-python==3.4.1.15
openpyxl==2.5.12
pandocfilters==1.4.2
parsel==1.4.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.2.0
prometheus-client==0.5.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycodestyle==2.4.0
pycparser==2.18
pycryptodome==3.7.3
PyDispatcher==2.0.5
Pygments==2.3.1
PyMySQL==0.8.1
pyOpenSSL==18.0.0
pyparsing==2.2.0
PyQt5==5.11.3
PyQt5-sip==4.19.13
python-dateutil==2.7.3
python-docx==0.8.7
python-editor==1.0.3
python-engineio==3.0.0
python-socketio==2.1.2
pytz==2018.4
PyYAML==4.2b1
pyzmq==17.1.2
qiniu==7.2.2
qtconsole==4.4.3
queuelib==1.5.0
redis==3.0.1
requests==2.18.4
resources==0.0.1
rsa==4.0
scipy==1.1.0
Scrapy==1.5.0
selenium==3.12.0
Send2Trash==1.5.0
service-identity==17.0.0
six==1.11.0
SQLAlchemy==1.2.14
staty==0.5.0
swagger-py-codegen==0.4.0
swagger-spec-validator==2.4.3
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
translate-toolkit==2.3.1
Twisted==18.4.0
typing==3.6.6
urllib3==1.22
uuid==1.30
virtualenv==16.1.0
virtualenvwrapper-win==1.2.5
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
xlrd==1.1.0
xlwt==1.3.0
yapf==0.22.0
